requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
torchaudio==0.8.0

-f https://dl.fbaipublicfiles.com/detectron2/wheels/cpu/torch1.8/index.html
detectron2==0.6

opencv-python-headless==4.6.0.66
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
pillow==9.3.0
pyyaml==6.0
requests==2.28.1
scipy==1.7.3
streamlit==1.15.1
tqdm==4.64.1
scikit-image==0.19.3